Implemented search results page using multi_match query and highlighting (no aggregations yet).
The multi_match query is safer than using the query DSL because it
does not go through a parsing process and has a much smaller chance of
failing. For now, this should ensure a reliable search experience for
users while working on other aspects of the search interface, such as
getting aggregations/faceting implemented. The downside of using
multi_match is it's not as powerful as the query DSL for end-users.

Perhaps the query DSL could be made available to super-users but
not regular uesrs. Something to think about.